# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the Biblical Verses app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VerseRequest {
  /** The user's problem or feeling described in text */
  problem: string;
  /** Optional list of verse IDs already shown (for no-repeat logic) */
  excluded_ids?: number[];
}

export interface VerseResponse {
  /** The detected emotional category */
  detected_category: string;
  /** A brief supportive message related to the category */
  message: string;
  /** The biblical reference (e.g. "Filipenses 4:6") */
  verse_reference: string;
  /** The full verse text */
  verse_text: string;
  /** The database ID of the returned verse */
  verse_id: number;
}

export interface Category {
  category: string;
  label: string;
  verse_count: number;
}

export interface CategoryStat {
  category: string;
  count: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetRandomVerseParams = {
  /**
   * Comma-separated list of verse IDs to exclude (already shown)
   */
  excluded_ids?: string;
};