# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for the Biblical Verses app
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: verses
    description: Biblical verse operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /verse:
    post:
      operationId: getVerse
      tags: [verses]
      summary: Get a biblical verse for a problem
      description: Analyzes the user's problem text, detects the emotional category, and returns a relevant biblical verse.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerseRequest"
      responses:
        "200":
          description: A biblical verse response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerseResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No verse found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      operationId: getCategories
      tags: [verses]
      summary: Get all available verse categories
      description: Returns a list of all available emotional categories.
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /verse/random:
    get:
      operationId: getRandomVerse
      tags: [verses]
      summary: Get a random verse
      description: Returns a random biblical verse from any category, optionally excluding already-seen verse IDs.
      parameters:
        - name: excluded_ids
          in: query
          required: false
          description: Comma-separated list of verse IDs to exclude (already shown)
          schema:
            type: string
      responses:
        "200":
          description: A random biblical verse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerseResponse"

  /verse/stats:
    get:
      operationId: getVerseStats
      tags: [verses]
      summary: Get verse statistics
      description: Returns counts of verses per category.
      responses:
        "200":
          description: Verse statistics by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryStat"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    VerseRequest:
      type: object
      properties:
        problem:
          type: string
          description: The user's problem or feeling described in text
        excluded_ids:
          type: array
          items:
            type: integer
          description: Optional list of verse IDs already shown (for no-repeat logic)
      required:
        - problem

    VerseResponse:
      type: object
      properties:
        detected_category:
          type: string
          description: The detected emotional category
        message:
          type: string
          description: A brief supportive message related to the category
        verse_reference:
          type: string
          description: The biblical reference (e.g. "Filipenses 4:6")
        verse_text:
          type: string
          description: The full verse text
        verse_id:
          type: integer
          description: The database ID of the returned verse
      required:
        - detected_category
        - message
        - verse_reference
        - verse_text
        - verse_id

    Category:
      type: object
      properties:
        category:
          type: string
        label:
          type: string
        verse_count:
          type: integer
      required:
        - category
        - label
        - verse_count

    CategoryStat:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required:
        - category
        - count

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error