# lib/api-zod/src/generated/types/verseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the Biblical Verses app
 * OpenAPI spec version: 0.1.0
 */

export interface VerseRequest {
  /** The user's problem or feeling described in text */
  problem: string;
  /** Optional list of verse IDs already shown (for no-repeat logic) */
  excluded_ids?: number[];
}