# lib/api-zod/src/generated/types/verseResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the Biblical Verses app
 * OpenAPI spec version: 0.1.0
 */

export interface VerseResponse {
  /** The detected emotional category */
  detected_category: string;
  /** A brief supportive message related to the category */
  message: string;
  /** The biblical reference (e.g. "Filipenses 4:6") */
  verse_reference: string;
  /** The full verse text */
  verse_text: string;
  /** The database ID of the returned verse */
  verse_id: number;
}